All EmailServer steps refactor
