+++ Projects/RMAutomationUIFramework/src/test/java/org/fundacionjala/automation/scenario/steps/admin/service/ServiceGivenSteps.java
@@ -1,20 +1,16 @@
 package org.fundacionjala.automation.scenario.steps.admin.service;
 
+import org.fundacionjala.automation.framework.pages.admin.conferencerooms.ConferenceRoomsPage;
 import org.fundacionjala.automation.framework.pages.admin.emailserver.AddEmailServerPage;
 import org.fundacionjala.automation.framework.pages.admin.emailserver.EmailServerPage;
 import org.fundacionjala.automation.framework.pages.admin.login.LoginPage;
 import org.fundacionjala.automation.framework.utils.common.BrowserManager;
-import org.testng.Assert;
 
 import cucumber.api.java.en.Given;
-import cucumber.api.java.en.Then;
-import cucumber.api.java.en.When;
 
-public class ServiceCredentialModified {
-	private String newUserName = "RoomManager1";
-	private String newPassword = "Control*123";
+public class ServiceGivenSteps {
 	
-	@Given("^There is a service existent$")
+	@Given("^There is an Email Server added$")
 	public void there_is_a_service_existent() throws Throwable {
 		BrowserManager.openBrowser();
 		LoginPage login = new LoginPage();
@@ -41,9 +37,10 @@ public class ServiceCredentialModified {
 					.clickSaveButton();
 		}
 	}
-
-	@When("^I modify credential with another account$")
-	public void i_modify_credential_with_another_account() throws Throwable {
+	
+	@Given("^There is no Email Server added$")
+	public void there_is_no_services_added() throws Throwable {
+		BrowserManager.openBrowser();
 		LoginPage login = new LoginPage();
 		
 		EmailServerPage emailServer = login
@@ -54,38 +51,53 @@ public class ServiceCredentialModified {
 				.leftMenu
 				.clickOnEmailServerButton();
 		
-		emailServer
-				.clickOnServerButton()
-				.clickOnEditCredentialButton()
-				.setUserName(newUserName)
-				.setPassword(newPassword)
-				.clickOnAcceptButton();
+		boolean isEmailServerPresent = emailServer.findEmailServer();
+		
+		if(isEmailServerPresent == true) {
+			
+			emailServer
+				.clickOnRemoveButton()
+				.clickOnYesButton();
+		}
 	}
-
-	@Then("^The changes are saved$")
-	public void the_changes_are_saved() throws Throwable {
-		EmailServerPage emailServer = new EmailServerPage();
-		
-		String myActualUserName = emailServer
-				.getUserName();
-		
-		Assert.assertEquals(myActualUserName, newUserName);
-		
+	
+	@Given("^At least a meeting$")
+	public void at_least_a_meeting() throws Throwable {
 		LoginPage login = new LoginPage();
 		
-		EmailServerPage server = login
-				.setUserName("SamuelSahonero")
-				.setPassword("Control*123")
-				.clickOnSigInButton()
-				.refreshPage()
-				.leftMenu
-				.clickOnEmailServerButton();
-		
-		server
-			.clickOnServerButton()
-			.clickOnEditCredentialButton()
-			.setUserName("Administrator")
+		ConferenceRoomsPage conferenceRooms = login
+			.setUserName("SamuelSahonero")
 			.setPassword("Control*123")
-			.clickOnAcceptButton();
+			.clickOnSigInButton()
+			.refreshPage()
+			.leftMenu
+			.clickOnIssuesButton()
+			.clickOnConferenceRoomsButton();
+		
+		conferenceRooms
+			.doubleClickOnRoom("Room04")
+			.SelectOutOfOrder()
+			.activeOutOfOrder()
+			.clickOnSave();
+	}
+	
+	@Given("^At least an out of order$")
+	public void at_least_an_out_of_order() throws Throwable {
+		LoginPage login = new LoginPage();
+		
+		ConferenceRoomsPage conferenceRooms = login
+			.setUserName("SamuelSahonero")
+			.setPassword("Control*123")
+			.clickOnSigInButton()
+			.refreshPage()
+			.leftMenu
+			.clickOnIssuesButton()
+			.clickOnConferenceRoomsButton();
+		
+		conferenceRooms
+			.doubleClickOnRoom("Room04")
+			.SelectOutOfOrder()
+			.activeOutOfOrder()
+			.clickOnSave();
 	}
 }
